changed virtual segment manager to correctly setup the centers
We are now representing for the calibration each segment as a center and
a radius, with a is_single bool to denotate if the center is just one
led, or between leds.

when it's between leds (is_single = false) the radius must be greater
than 1, when it's a single led radius can be 0.

The length is given by `(radius * 2) + 1 * is_single`.

First start with radius = 1 for all the leds and center each one. Then
start increasing the radius till it reaches the border at the same time
from both sides. If that's not the case, correct the center.